handle data flow board using redux, usedispatch, useselector
